fix: [CDS-81585]: Fix for Integer Overflow in case of auto approval timeout  (#55018)
* fix: [CDS-81585]: have converted int32 to long

* fix: [CDS-81585]: rebase

* fix: add unit to test for timeout greater than int32

+++ clients/pipeline-service/pms-sdk-core/src/main/java/io/harness/pms/sdk/core/waiter/AsyncWaitEngine.java
@@ -17,7 +17,7 @@ import java.util.List;
 @OwnedBy(HarnessTeam.PIPELINE)
 public interface AsyncWaitEngine {
   void waitForAllOn(
-      NotifyCallback notifyCallback, ProgressCallback progressCallback, List<String> correlationIds, int timeout);
+      NotifyCallback notifyCallback, ProgressCallback progressCallback, List<String> correlationIds, long timeout);
 
   /**
    * This method notifies that they have acquired the task for the given correlationId
